代码检视修改： bHasClosed_ 删除、用fd_去判断连接状态 bHasClosed_ 删除、用fd_去判断连接状态，fd_ != -1 设置fd_=-1 这个线程是否需要detach掉，join是不是更合适 bool OHOS::MMI::ServerMsgHandler::Init(UDSServer& udsServer) 返回值改为void Start 客户端启动失败需要输出日志 EventManager.SetClientHandle(mClient_); 这里和 Start中的 EventManager.SetClientHandle(GetPtr()); 代码重复，删除一个 研究buf.size()能不能改成buf.length() int32_t size[1]; 研究改为 int32_t size;

+++ frameworks/util/socket/src/uds_client.cpp
@@ -149,7 +149,7 @@ void UDSClient::OnRecv(const char *buf, size_t size)
             return;
         }
         auto head = reinterpret_cast<PackHead *>(const_cast<char *>(&buf[readIdx]));
-        if (head->size < 0 || head->size >= static_cast<int32_t>(bufSize)) {
+        if (head->size < 0 || head->size >= bufSize) {
             MMI_LOGE("Head size is error, head->size:%{public}d, errCode:%{public}d", head->size, VAL_NOT_EXP);
             return;
         }
